更新优化 nginx

+++ docker/sub-store.sh
@@ -19,8 +19,9 @@ echo -e "==============================================================\n"
 
 echo -e "======================== 2. 启动nginx ========================\n"
 
-    echo -e "生成 nginx 配置文件\n"
-    envsubst '${ALLOW_IP}' < /etc/nginx/conf.d/front.template > /etc/nginx/conf.d/front.conf
+    if [ ! -f "/etc/nginx/conf.d/front.conf" ]; then
+        echo -e "生成 nginx 配置文件\n"
+        envsubst '${ALLOW_IP}' < /etc/nginx/conf.d/front.template > /etc/nginx/conf.d/front.conf    
     nginx -s reload 2>/dev/null || nginx -c /etc/nginx/nginx.conf
 
 echo -e "==============================================================\n"
